+++ src/NetCore/Web/Api/wwwroot/scripts/api/webgis.js
@@ -2633,7 +2633,7 @@
                     webgis.showFeatureResultModal(map.guid, feature.oid, queryToolId);
                 } else {
                     var $currentTabContent = $tabControl.webgis_tab_control('currentContent');
-                    $currentTabContent.webgis_queryResultsTable('selectRow', { map: map, dataId: feature.oid, scrollTo: true });
+                    $currentTabContent.webgis_queryResultsTable('selectRow', { map: map, dataId: feature.oid, scrollTo: e.suppressScrollResultRowTo !== true });
                 }
                 map.queryResultFeatures.tryHighlightMarker(feature.oid, true);